br-01-keys
Elements within a collection must be provided with keys in order to have a stable identity. React uses keys to determine which elements in the collection need to be re-created and re-rendered rather than using elements from a previous render. This way we avoid recreating all the elements of the collection every time something changes.